+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Upcoming
+### Fix
+
+* Issue #2: Escaping of promise return value is now done properly 
+
 ## v0.2.0 - 2015-07-26
 ### Change
 * `registerHelper` is not overwritten anymore. Instead, a new method `registerPromiseHelper`